Update patch set 3
Patch Set 3:

I can't speak for the team... I don't really do much around here. But the conclusion I have come to is that if we are going to workaround complying with CTS in the kernel, then we might as well do it here and give users the choice on whether or not they want to meet CTS compliance. As for getting blacklisted by Google, I don't see that as a possibility. We are a custom ROM for a few devices. If we are blacklisted, then at the very least Google will have improved their security as they have done through other open platforms such as Magisk. But again, this is not my decision. I don't call the shots here

Patch-set: 3
Reviewer: Gerrit User 1000166 <1000166@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Code-Review=0