Update patch set 1
Patch Set 1:

In my testing on the Razer Phone, this was a regression on 8998 compared to A9 routines. Qualcomm's testing on O seems to agree based on their commit messages. Not sure if it really is a win for pixel 2. If it is, it may be because of scheduling differences between QC HMP and EAS. Just something to think about for anyone maintaining an 8998 device. Always do regression testing even if it's from Google.

Patch-set: 1
Reviewer: Gerrit User 1000012 <1000012@3162e3d8-d894-470e-abd9-724f7b0effad>
Label: Verified=0