Update patch set 1
Patch Set 1:

I suggest we remove this functionality completely. It seems to just be unnecessary complication on the part of CAF to move things to a proprietary blob. The easy way to remove it is to remove the -DUSE_WRAPPER flags in libc/Android.bp, or revert commit a707cc7fd2979918f4ff69c02c5a94e4bbc2401a.

Patch-set: 1